+++ source-files/web-components/notice-panel.js
@@ -56,11 +56,9 @@ window.customElements.define('notice-panel', class extends HTMLElement {
     panelMessageDisplayEl.value +=
       document.getElementById('displayUtils').cleanFormatting(ctcpMessage) + '\n';
 
-    if (!window.globals.webState.cacheReloadInProgress) {
-      if ((!('zoomPanelId' in window.globals.webState)) ||
-        (window.globals.webState.zoomPanelId.length < 1)) {
-        this.showPanel();
-      }
+    if ((!window.globals.webState.cacheReloadInProgress) &&
+      (!document.querySelector('body').hasAttribute('zoomId'))) {
+      this.showPanel();
     }
 
     if (!window.globals.webState.cacheReloadInProgress) {
@@ -118,11 +116,9 @@ window.customElements.define('notice-panel', class extends HTMLElement {
       return;
     };
 
-    if (!window.globals.webState.cacheReloadInProgress) {
-      if ((!('zoomPanelId' in window.globals.webState)) ||
-        (window.globals.webState.zoomPanelId.length < 1)) {
-        this.showPanel();
-      }
+    if ((!window.globals.webState.cacheReloadInProgress) &&
+      (!document.querySelector('body').hasAttribute('zoomId'))) {
+      this.showPanel();
     }
 
     if (panelDivEl.getAttribute('lastDate') !== parsedMessage.datestamp) {
@@ -356,10 +352,10 @@ window.customElements.define('notice-panel', class extends HTMLElement {
      * @listens document:resize-custom-elements
      */
     document.addEventListener('resize-custom-elements', () => {
-      if (window.globals.webState.dynamic.inputAreaCharWidthPx) {
+      if (window.globals.webState.dynamic.testAreaColumnPxWidth) {
         const calcInputAreaColSize = document.getElementById('displayUtils').calcInputAreaColSize;
         // pixel width mar1 is reserved space on edges of input area at full screen width
-        const mar1 = window.globals.webState.dynamic.commonMargin;
+        const mar1 = window.globals.webState.dynamic.commonMarginRightPx;
         // set width of input area elements
         this.shadowRoot.getElementById('panelMessageDisplayId')
           .setAttribute('cols', calcInputAreaColSize(mar1));