+++ source-files/web-components/show-raw.js
@@ -200,10 +200,10 @@ window.customElements.define('show-raw', class extends HTMLElement {
      * @listens document:resize-custom-elements
      */
     document.addEventListener('resize-custom-elements', () => {
-      if (window.globals.webState.dynamic.inputAreaCharWidthPx) {
+      if (window.globals.webState.dynamic.testAreaColumnPxWidth) {
         const calcInputAreaColSize = document.getElementById('displayUtils').calcInputAreaColSize;
         // pixel width mar1 is reserved space on edges of input area at full screen width
-        const mar1 = window.globals.webState.dynamic.commonMargin;
+        const mar1 = window.globals.webState.dynamic.commonMarginRightPx;
         // set width of input area elements
         this.shadowRoot.getElementById('panelMessageDisplayId')
           .setAttribute('cols', calcInputAreaColSize(mar1));